feat: create index component to encapsulate the main functionality form the app component
